refactor: Move ClubInfoWrapper component to /club/layout.tsx
